Added boolean argument delete_installer to download_pandoc

+++ pypandoc/pandoc_download.py
@@ -149,7 +149,7 @@ def _handle_win32(filename, targetfolder):
     print("* Done.")
 
 
-def download_pandoc(url=None, targetfolder=None, version="latest"):
+def download_pandoc(url=None, targetfolder=None, version="latest", delete_installer=False):
     """Download and unpack pandoc
 
     Downloads prebuild binaries for pandoc from `url` and unpacks it into
@@ -205,4 +205,5 @@ def download_pandoc(url=None, targetfolder=None, version="latest"):
     assert unpack is not None, "Can't handle download, only Linux, Windows and OS X are supported."
 
     unpack(filename, targetfolder)
-    os.remove(filename)
+    if delete_installer:
+        os.remove(filename)
